requirements.txt
djangorestframework
# This file is expected by Heroku.

-r requirements/production.txt
